Rename register_translation to _register_translation
