+++ skimage/registration/__init__.py
@@ -1,5 +1,5 @@
 from ._optical_flow import optical_flow_tvl1
-from .register_translation import register_translation
+from ._register_translation import register_translation
 
 
 __all__ = [
