+++ skimage/registration/_register_translation.py
